latest update

+++ demo/src/main/java/com/example/demo/repo/PatientRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.Patient;
 
 @Repository
-public interface PatientRepo extends JpaRepository<Patient , Integer>{
+public interface PatientRepo extends JpaRepository<Patient , Long>{
     Patient findByUsername(String Username);
     Patient findByUsernameAndTyp(String username,String typ);
     Patient findByEmailAndTyp(String email,String typ);
